refactor(web): wrap recordings overview and mode views in LoadedFade
- Refactored `RecordingsOverview` to use `LoadedFade` for improved UI loading animations.
- Updated imports across various routes and components to utilize the renamed `LoadedFade` component.
- Introduced `modeViewKey` to ensure re-mounting of main view content for enhanced fade animation on mode/view changes.
