+++ web/src/lib/components/youtube/index.ts
@@ -1,3 +1,3 @@
-export { default as LoadedFade } from "./LoadedFade.svelte";
+export { default as LoadedFade } from "../LoadedFade.svelte";
 export { default as YouTubeListState } from "./YouTubeListState.svelte";
 export { default as YouTubeMediaRow } from "./YouTubeMediaRow.svelte";
